cmake error

+++ app/pidcont.cpp
@@ -14,10 +14,9 @@ pidcont::~pidcont(){
 
 }
 
-float pidcont::computenew_velo(double targetvelo,double actuvelo double Kp_input, double Ki_input, double Kd_input) {
-
-double someconst = 3.1;
-return someconst;  
+double pidcont::computenew_velo(double targetvelo, double actuvelo, double Kp_input, double Ki_input, double Kd_input) {
+  double someconst = 3.1;
+  return someconst;  
 }
